requirements.txt
pandas
requests
openpyxl
boto3
numpy
python-dotenv
pyarrow